Some minor changes necessary for PGCC4.
git-svn-id: https://svn.dealii.org/trunk@6743 0785d39b-7218-0410-832d-ea1e28bc413d

+++ deal.II/base/source/polynomial.cc
@@ -239,7 +239,8 @@ namespace Polynomials
   {
     for (int i=degree();i>=0;--i)
       {
-        out << coefficients[i] << " x^" << i << std::endl;
+        out << static_cast<double>(coefficients[i])
+            << " x^" << i << std::endl;
       }
   }
